+++ react/typings/vtex.react-vtexid.d.ts
@@ -1,0 +1,7 @@
+/* Typings for `react-vtexid` */
+declare module 'vtex.react-vtexid' {
+  import { Component, ComponentType, ReactElement, ReactType } from 'react'
+
+  export const useRedirectLogout: any
+
+}